Paladin stake integration

+++ contracts/interfaces/external/paladin/IPaladinZap.sol
@@ -1,15 +1,13 @@
 pragma solidity ^0.7.6;
 
 interface IPaladinZap {
-
-  function zapDeposit(
-      address _fromTokenAddress,
-      address _toTokenAddress,
-      address _poolAddress,
-      uint256 _amount,
-      address _swapTarget,
-      address _allowanceTarget,
-      bytes memory _swapData
-  ) external payable returns(uint);
-
+    function zapDeposit(
+        address _fromTokenAddress,
+        address _toTokenAddress,
+        address _poolAddress,
+        uint256 _amount,
+        address _swapTarget,
+        address _allowanceTarget,
+        bytes memory _swapData
+    ) external payable returns (uint256);
 }
